Move scan_for_imports pyfunction into import_scanning module
